# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sarkari Exams API",
    "version": "1.0.0",
    "description": "Shared API contract for frontend/backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/community/flags": {
      "post": {
        "summary": "Report abusive content",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityFlagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Flag",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityFlagResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/community/forums": {
      "get": {
        "summary": "List forum posts",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forum posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityForumListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create forum post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityForumCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Forum post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityForumResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/community/groups": {
      "get": {
        "summary": "List study groups",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Study groups",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityGroupListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create study group",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityGroupCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Study group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityGroupResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/community/qa": {
      "get": {
        "summary": "List Q&A threads",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Q&A threads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityQaListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Q&A thread",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityQaCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Q&A thread",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityQaResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/push/subscribe": {
      "post": {
        "summary": "POST /api/push/subscribe",
        "tags": [
          "push"
        ],
        "operationId": "postApiPushSubscribe",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/push/vapid-public-key": {
      "get": {
        "summary": "GET /api/push/vapid-public-key",
        "tags": [
          "push"
        ],
        "operationId": "getApiPushVapidPublicKey",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommunityEntityType": {
        "type": "string",
        "enum": [
          "forum",
          "qa",
          "group"
        ]
      },
      "CommunityFlagStatus": {
        "type": "string",
        "enum": [
          "open",
          "reviewed",
          "resolved"
        ]
      },
      "CommunityForumPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "category",
          "author",
          "createdAt",
          "updatedAt"
        ]
      },
      "CommunityQaThread": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string",
            "nullable": true
          },
          "answeredBy": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "question",
          "author",
          "createdAt",
          "updatedAt"
        ]
      },
      "CommunityStudyGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "link": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "topic",
          "language",
          "createdAt",
          "updatedAt"
        ]
      },
      "CommunityFlag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "entityType": {
            "$ref": "#/components/schemas/CommunityEntityType"
          },
          "entityId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "reporter": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/CommunityFlagStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "entityType",
          "entityId",
          "reason",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CommunityForumCreateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content",
          "category"
        ]
      },
      "CommunityQaCreateRequest": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        },
        "required": [
          "question"
        ]
      },
...
      "CommunityGroupCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "link": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "topic",
          "language"
        ]
      },
      "CommunityFlagCreateRequest": {
        "type": "object",
        "properties": {
          "entityType": {
            "$ref": "#/components/schemas/CommunityEntityType"
          },
          "entityId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "reporter": {
            "type": "string"
          }
        },
        "required": [
          "entityType",
          "entityId",
          "reason"
        ]
      },
      "CommunityForumResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CommunityForumPost"
          }
        },
        "required": [
          "data"
        ]
      },
      "CommunityQaResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CommunityQaThread"
          }
        },
        "required": [
          "data"
        ]
      },
      "CommunityGroupResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CommunityStudyGroup"
          }
        },
        "required": [
          "data"
        ]
      },
      "CommunityFlagResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CommunityFlag"
          }
        },
        "required": [
          "data"
        ]
      },
      "CommunityForumListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommunityForumPost"
            }
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "data"
        ]
      },
      "CommunityQaListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommunityQaThread"
            }
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "data"
        ]
      },
      "CommunityGroupListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommunityStudyGroup"
            }
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}